requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
anthropic>=0.40.0
tenacity>=8.5.0
PyYAML>=6.0.1
